gis_gkh_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-lru==2.0.4
async-property==0.2.2
async-timeout==4.0.3
asyncio-throttle==1.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
dadata==21.10.1
et-xmlfile==1.1.0
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
gspread==5.10.0
gspread-asyncio==1.9.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
jmespath==1.0.1
loguru==0.7.2
lxml==4.9.4
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.4
multidict==6.0.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
playwright==1.40.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.0.1
pygsheets==2.0.6
pyparsing==3.1.1
pypdf==3.17.4
PySocks==1.7.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.9.0
undetected-chromedriver==3.5.4
uritemplate==4.1.1
urllib3==2.1.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
